requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery
redis
psycopg2-binary
anthropic
python-dotenv
Pillow
pytesseract
channels
channels-redis
drf-spectacular
